Trim itfMesManager loop locals

+++ src/itfMesManager.c
@@ -1512,11 +1512,11 @@ void FUN_003a3530(u32 *param_1)
 
   int iVar3;
 
-  long lVar4;
+  int lVar4;
 
-  long lVar5;
+  int lVar5;
 
-  long lVar6;
+  int lVar6;
